# openapi.yaml
openapi: 3.0.0
info:
  title: 车辆电池监控系统 API
  description: 车辆电池监控系统的API接口文档
  version: 1.0.0
  contact:
    name: 系统管理员
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
...
paths:
...
  /api/warn:
    post:
      tags:
        - reportwarn
      summary: 处理电池信号并返回预警
      description: 上报电池信号并根据预警规则生成预警信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BatterySignalUploadReq'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatteryWarnRespListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/report:
    post:
      tags:
        - reportwarn
      summary: 上报电池信号
      description: 批量上报电池信号信息，不触发预警处理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BatterySignalUploadReq'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatterySignalListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/search:
    get:
      tags:
        - reportwarn
      summary: 查询预警信息
      description: 根据条件查询车辆的预警信息
      parameters:
        - name: car_id
          in: query
          description: 车辆ID
          required: true
          schema:
            type: integer
        - name: warn_id
          in: query
          description: 预警ID
          required: false
          schema:
            type: integer
        - name: min_warn_time
          in: query
          description: 最小预警时间
          required: false
          schema:
            type: string
            format: date-time
        - name: max_warn_time
          in: query
          description: 最大预警时间
          required: false
          schema:
            type: string
            format: date-time
        - name: handled
          in: query
          description: 是否已处理
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarnInfoListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: 车辆不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码
          example: 200
        message:
          type: string
          description: 状态消息
          example: "success"
        data:
          description: 返回数据
          nullable: true
      required:
        - code
        - message
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: 'null'
      example:
        code: 400
        message: "参数错误"
        data: null
...
    BatterySignal:
      type: object
      properties:
        id:
          type: integer
          description: 主键ID
        carId:
          type: integer
          description: 车辆ID
        warnId:
          type: integer
          description: 预警标识
        signalData:
          type: string
          description: 信号数据（JSON格式）
        signalTime:
          type: string
          format: date-time
          description: 信号时间
        processed:
          type: boolean
          description: 是否已处理
        processTime:
          type: string
          format: date-time
          description: 处理时间
        createTime:
          type: string
          format: date-time
          description: 创建时间
...
    BatterySignalListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BatterySignal'
    
    # 预警规则相关模型
...
    WarnInfo:
      type: object
      properties:
        id:
          type: integer
          description: 主键ID
        carId:
          type: integer
          description: 车辆ID
        warnId:
          type: integer
          description: 预警ID
        signalId:
          type: integer
          format: int64
          description: 信号ID
        warnTime:
          type: string
          format: date-time
          description: 预警时间
        warnLevel:
          type: integer
          description: 预警等级
        warnDesc:
          type: string
          description: 预警描述
        handled:
          type: boolean
          description: 是否已处理
        handleTime:
          type: string
          format: date-time
          description: 处理时间
        createTime:
          type: string
          format: date-time
          description: 创建时间
...
    WarnInfoListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WarnInfo'
    
    # 电池信号上传请求
    BatterySignalUploadReq:
      type: object
      properties:
        carId:
          type: integer
          description: 车架编号
        warnId:
          type: integer
          description: 预警编号，如1表示电压差检测，2表示电流差检测。若不提供，则对所有预警规则进行检查
          minimum: 1
        signal:
          type: object
          description: 信号数据，包含电池相关指标
        vid:
          type: string
          description: 车辆VID，用于双重验证
          pattern: '^[A-Z0-9]{16}$'
        signalTime:
          type: string
          format: date-time
          description: 信号时间，不提供则使用系统当前时间
      required:
        - carId
        - signal
    
    # 电池预警响应
    BatteryWarnResp:
      type: object
      properties:
        signal:
          $ref: '#/components/schemas/BatterySignal'
        parsedSignal:
          type: object
          description: 解析后的信号数据
        warnings:
          type: array
          description: 生成的预警信息
          items:
            $ref: '#/components/schemas/WarnInfo'
    BatteryWarnRespListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BatteryWarnResp'
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            message: "参数校验不通过：carId: 车架编号不能为空"
            data: null
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: "系统异常：数据库连接失败"
            data: null 